refactor: useCalcWidthRef 커스텀훅 분리
