+++ src/lib/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useIsMobile } from './useIsMobile';
 export { useIsTablet } from './useIsTablet';
 export { useIsMounted } from './useIsMounted';
+export { useCalcWidthRef } from './useCalcWidthRef';
